Versuch, den Shader einzufügen

+++ Computergraphik/Praktikum06/Blatt01/Model.h
@@ -56,7 +56,7 @@ private:
 	GLenum mode;
 
 	glm::vec3 material;
-	int shininess;
+	float shininess;
 
 	bool active;
 };
